refactor: #747 MyProfileUiState, MyProfileUiEvent 패키지 이동
