+++ android/TodokTodok/app/src/androidTest/java/com/team/todoktodok/compose/my/participated/ParticipatedDiscussionsScreenScreenTest.kt
@@ -31,8 +31,8 @@ class ParticipatedDiscussionsScreenScreenTest {
                             createAt = LocalDateTime.of(2025, 7, 12, 12, 0),
                             discussionOpinion =
                                 "응집도와 결합도가 어떤 차이를 가지는 지에 대한 다른 분들의 생각이 궁금합니다." +
-                                        "응집도는 내부에 얼마나 비슷한 책임들이 모여있는가. 얼마나 연관있는 멤버들이 똘똘 뭉쳐있는가" +
-                                        "응집도는 내부에 얼마나 비슷한 책임들이 모여있는가. 얼마나 연관있는 멤버들이 똘똘 뭉쳐있는가",
+                                    "응집도는 내부에 얼마나 비슷한 책임들이 모여있는가. 얼마나 연관있는 멤버들이 똘똘 뭉쳐있는가" +
+                                    "응집도는 내부에 얼마나 비슷한 책임들이 모여있는가. 얼마나 연관있는 멤버들이 똘똘 뭉쳐있는가",
                             likeCount = 0,
                             commentCount = 0,
                             viewCount = 0,
@@ -80,7 +80,8 @@ class ParticipatedDiscussionsScreenScreenTest {
                 uiModel = uiModel,
                 onChangeShowMyDiscussion = {},
                 onCompleteRemoveDiscussion = {},
-                onCompleteModifyDiscussion = {})
+                onCompleteModifyDiscussion = {},
+            )
         }
 
         // Then
@@ -98,7 +99,8 @@ class ParticipatedDiscussionsScreenScreenTest {
                 uiModel = testUiModel,
                 onChangeShowMyDiscussion = {},
                 onCompleteRemoveDiscussion = {},
-                onCompleteModifyDiscussion = {})
+                onCompleteModifyDiscussion = {},
+            )
         }
 
         // Then
@@ -118,7 +120,8 @@ class ParticipatedDiscussionsScreenScreenTest {
                 uiModel = testUiModel,
                 onChangeShowMyDiscussion = {},
                 onCompleteRemoveDiscussion = {},
-                onCompleteModifyDiscussion = {})
+                onCompleteModifyDiscussion = {},
+            )
         }
 
         // Then
@@ -136,7 +139,8 @@ class ParticipatedDiscussionsScreenScreenTest {
                 uiModel = testUiModel,
                 onChangeShowMyDiscussion = {},
                 onCompleteRemoveDiscussion = {},
-                onCompleteModifyDiscussion = {})
+                onCompleteModifyDiscussion = {},
+            )
         }
 
         // Then
