+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/preview/MyProfileUiStatePreviewParameterProvider.kt
@@ -2,7 +2,7 @@ package com.team.todoktodok.presentation.compose.preview
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import com.team.domain.model.member.Profile
-import com.team.todoktodok.presentation.compose.my.model.MyProfileUiState
+import com.team.todoktodok.presentation.compose.my.MyProfileUiState
 
 class MyProfileUiStatePreviewParameterProvider : PreviewParameterProvider<MyProfileUiState> {
     override val values: Sequence<MyProfileUiState>
